Working cart page, localStorage, modularized components-Ahmed
