+++ src/js/loadComponent.js
@@ -3,7 +3,7 @@ export async function loadComponent(componentPath, targetId) {
       const response = await fetch(componentPath);
       const content = await response.text();
       document.getElementById(targetId).innerHTML = content;
-    } catch (error) {
+    } catch (error) {t
       console.error(`Error loading ${componentPath}:`, error);
     }
   }
